AuthScreen [nfc]: Fold in activeAuthentications and its data.
This is super closely tied to the AuthScreen component: witness
especially the `handler` values, which are interpreted as names of
methods on the component (yuck!).

So, move it into the same module -- the better to be read and
understood together, and to evolve together.
